parsing/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-googlesearch==0.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
google-search-results==2.4.2
googlesearch-python==1.2.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
newsdataapi==0.1.6
openai==0.27.4
protobuf==3.20.3
requests==2.28.2
soupsieve==2.4.1
SQLAlchemy==2.0.9
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
